Created RequestModule
