+++ scripts/system/makeUserConnection.js
@@ -9,6 +9,9 @@
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
+
+var RequestModule = Script.require('./request.js');
+
 (function() { // BEGIN LOCAL_SCOPE
 
     var LABEL = "makeUserConnection";
@@ -121,63 +124,12 @@
             [].map.call(arguments, JSON.stringify)));
     }
 
+    function debugPrint() {
+        print( [].slice.call(arguments).join(' ') );
+    }
+
     function cleanId(guidWithCurlyBraces) {
         return guidWithCurlyBraces.slice(1, -1);
-    }
-    function request(options, callback) { // cb(error, responseOfCorrectContentType) of url. A subset of npm request.
-        var httpRequest = new XMLHttpRequest(), key;
-        // QT bug: apparently doesn't handle onload. Workaround using readyState.
-        httpRequest.onreadystatechange = function () {
-            var READY_STATE_DONE = 4;
-            var HTTP_OK = 200;
-            if (httpRequest.readyState >= READY_STATE_DONE) {
-                var error = (httpRequest.status !== HTTP_OK) && httpRequest.status.toString() + ':' + httpRequest.statusText,
-                    response = !error && httpRequest.responseText,
-                    contentType = !error && httpRequest.getResponseHeader('content-type');
-                if (!error && contentType.indexOf('application/json') === 0) { // ignoring charset, etc.
-                    try {
-                        response = JSON.parse(response);
-                    } catch (e) {
-                        error = e;
-                    }
-                }
-                if (error) {
-                    response = {statusCode: httpRequest.status};
-                }
-                callback(error, response);
-            }
-        };
-        if (typeof options === 'string') {
-            options = {uri: options};
-        }
-        if (options.url) {
-            options.uri = options.url;
-        }
-        if (!options.method) {
-            options.method = 'GET';
-        }
-        if (options.body && (options.method === 'GET')) { // add query parameters
-            var params = [], appender = (-1 === options.uri.search('?')) ? '?' : '&';
-            for (key in options.body) {
-                if (options.body.hasOwnProperty(key)) {
-                    params.push(key + '=' + options.body[key]);
-                }
-            }
-            options.uri += appender + params.join('&');
-            delete options.body;
-        }
-        if (options.json) {
-            options.headers = options.headers || {};
-            options.headers["Content-type"] = "application/json";
-            options.body = JSON.stringify(options.body);
-        }
-        for (key in options.headers || {}) {
-            if (options.headers.hasOwnProperty(key)) {
-                httpRequest.setRequestHeader(key, options.headers[key]);
-            }
-        }
-        httpRequest.open(options.method, options.uri, true);
-        httpRequest.send(options.body);
     }
 
     function handToString(hand) {
@@ -420,6 +372,7 @@
     // initiate the shake, they will race to see who sends the connectionRequest after noticing the
     // waiting message.  Either way, they will start connecting eachother at that point.
     function startHandshake(fromKeyboard) {
+
         if (fromKeyboard) {
             debug("adding animation");
             // just in case order of press/unpress is broken
@@ -492,7 +445,7 @@
         }
         // No-op if we were successful, but this way we ensure that failures and abandoned handshakes don't leave us
         // in a weird state.
-        request({uri: requestUrl, method: 'DELETE'}, debug);
+        RequestModule.request({ uri: requestUrl, method: 'DELETE' }, debug);
     }
 
     function updateTriggers(value, fromKeyboard, hand) {
@@ -606,7 +559,7 @@
                 connectionRequestCompleted();
             } else { // poll
                 Script.setTimeout(function () {
-                    request({
+                    RequestModule.request({
                         uri: requestUrl,
                         // N.B.: server gives bad request if we specify json content type, so don't do that.
                         body: requestBody
@@ -658,7 +611,7 @@
 
         // This will immediately set response if successful (e.g., the other guy got his request in first),
         // or immediate failure, and will otherwise poll (using the requestBody we just set).
-        request({ //
+        RequestModule.request({ //
             uri: requestUrl,
             method: 'POST',
             json: true,
